Optimizations, fixes and so on
So.... In this commit I fixed all GUI issues (it was loading 2 times
before). Also, I changed method of how LoadContent is called. Now, in
your game screen, you msut call it manually, so you have more freedom in
passing properties and so to assets or building sprites from oaded
assets, using fonts and so. Also, added all points totile Object, so you
are not restricted to use built in physics to use objects correctly (but
it is reccomenned to use builtin physics, becouse it is epic :D). And I
did also minor changes to EntityList and big changes to repository
structure, but they are not important, just for easier navigation ;)